Enable database filter for genre and region
Add the missing items so that database manager (under main menu
Information) can filter also on these properties. There are many
others that are defined in the structure, but do not allow for this,
however, after a brief check they are not used at all in any current
database (such as "narrative" or "gameplay").