Add support for extended controller matching using a physical identifier
Extend the matching logic to cover also the cases when vid/pid is the same,
but still there is some difference between gamepads. These identifiers are
highly environment specific and are only meant to be edited locally, not
added to any central repository.

The controller affinity will now check a "phys" attribute as well. The value
of this attribute can be determined from RetroArch debug logs, and can be
added to the autoconfig file for better matching.

Currently, udev driver will store USB port and serial identifier here,
retrieved by EVIOCGPHYS and EVIOCGUNIQ.

Co-authored-by: Luther7718 <luther7718@gmail.com>