Scanning optimization
Extend scanning logic by filtering databases, so that queries that
will certainly be unsuccessful, are skipped.

Filtering criteria:
- size of scanned file
- presence of serial (in case of serials)

Two extra queries for min and max are run for each database on the
first scan, this causes single-file scans to be somewhat slower,
but it pays off with large scans.

Archive size is also checked. The built-in shortcut of scanning
both the archive (zip file) and the first (possibly only) file
inside the archive is retained, it makes logic more complicated,
but the speed difference is significant for sets that contain
one zipped file per title.